navbar behaves perfectly
So the navbar is collapsable. Accessibility is considered with "aria-" classes in the navbar. Moved the photos into the navbar as well. Colors are redone when collapsed versus expanded as well. (:

Cleaned up by taking out unused things, commented out things that may be useful in the future in some way.