navbar is completely done (for now)
hamburger menu icon is added and works!

TODO: if you click the ham icon too fast while the animation of the menu is being hidden or shown, it can make it so the ham icon is flipped the wrong way. Make it wait until that animation, so like, it has to be at least .25 seconds between each click.